+++ src/main/java/net/MrGise/mmm/event/ModBusClientEvents.java
@@ -1,10 +1,13 @@
 package net.MrGise.mmm.event;
 
 import net.MrGise.mmm.MMM;
+import net.MrGise.mmm.block.entity.renderer.ThingamajigBlockEntityRenderer;
 import net.MrGise.mmm.particle.ConfettiParticle;
 import net.MrGise.mmm.particle.ManaParticle;
+import net.MrGise.mmm.registry.content.ModBlockEntities;
 import net.MrGise.mmm.registry.decorative.ModParticles;
 import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.client.event.EntityRenderersEvent;
 import net.minecraftforge.client.event.RegisterParticleProvidersEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
@@ -21,4 +24,9 @@ public class ModBusClientEvents {
         event.registerSpriteSet(ModParticles.CONFETTI.get(), ConfettiParticle.GravProvider::new);
     }
 
+    @SubscribeEvent
+    public static void registerBER(EntityRenderersEvent.RegisterRenderers event) {
+        event.registerBlockEntityRenderer(ModBlockEntities.THINGAMAJIG_BE.get(),
+                ThingamajigBlockEntityRenderer::new);
+    }
 }