+++ .github/logs/CHANGELOG.md
@@ -36,6 +36,7 @@ Branch: Course branch
 - 14/5/2026: Added Confetti for my birthday, which is On the same day as Minecraft's!
 - 15/5/2026: Finished splitting project
 - 21/5/2026: Added fluid handling to Thingamajig
+- 23/6/2026: Added Rendering to Thingamajig block entity
 
 ## [0.4.1a] (skipping the -1.20.1 because it's not important)
 For the changelog summary, visit [.github/logs](../logs/0.4.1a-1.20.1.md)